requirements.txt
alembic==1.16.4
annotated-types==0.7.0
anyio==4.10.0
asyncpg==0.30.0
bcrypt==4.0.1
cffi==1.17.1
click==8.2.1
cryptography==45.0.6
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.116.1
greenlet==3.2.4
h11==0.16.0
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
mypy==1.17.1
mypy_extensions==1.1.0
passlib==1.7.4
pathspec==0.12.1
pyaes==1.6.1
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyJWT==2.10.1
PySocks==1.7.1
python-dotenv==1.1.1
ruff==0.12.9
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.2
types-passlib==1.7.7.20250602
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
